Add rating service, controller and model
This commit introduces the addition of a new rating domain. The domain consists of a service, controller and a model for managing ratings. The main functionality provided includes setting a rating for a specific model and retrieval of ratings for a given question ID. The question domain has also been updated to include the rating service to aggregate and return ratings for each question when listing or retrieving individual questions. Additionally, routes have also been updated to handle requests for the rating domain. This development provides the application with an expanded functionality and a better organization of code by separating concerns for rating into its own domain.